backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.4.0
celery==5.3.6
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
django-cors-headers==4.3.1
pyjwt==2.8.0
drf-spectacular==0.27.1
drf_yasg==1.21.8
django-prometheus==2.3.1
django-redis==5.4.0
openai>=1.0.0